requirements.txt
torch==1.13.1
torchvision==0.14.1
numpy==1.19.0
tqdm==4.65.0
tensorboard==2.12.2
sklearn==0.0
Pillow==9.0.1